+++ python/gramsci/__init__.py
@@ -1,4 +1,4 @@
-from ._core import compute_2pcf, compute_3pcf, compute_4pcf
+from ._core import compute_2pcf, compute_3pcf, compute_4pcf, hartlap_factor
 
-__all__ = ['compute_2pcf', 'compute_3pcf', 'compute_4pcf']
+__all__ = ['compute_2pcf', 'compute_3pcf', 'compute_4pcf', 'hartlap_factor']
 __version__ = '0.1.0'